+++ Lab_2/lab2/src/main/java/edu/nikita/lab2/Square.java
@@ -2,7 +2,7 @@ package edu.nikita.lab2;
 
 public class Square implements IFunction {
 
-	public Integer Compute(Integer n) throws Exception {
+	public Long Compute(Long n) throws Exception {
 		return n * n;
 	}
